Update the Emacs editor support link (#126)
We created and maintain a new package for gdscript support in Emacs, that's now on the melpa archive (plugins that can be installed straight from Emacs). Franco, the author of the previous package, already contributed to it too, along with some other persons. Our new mode is up to date, has more features, is supported in Doom Emacs and Spacemacs...